+++ Assets/RBSocket/Message/DefaultMsgs/std_msgs/UInt8.cs
@@ -5,7 +5,7 @@ namespace RBS.Messages.std_msgs
     [System.Serializable]
     public class UInt8 : ExtendMessage
     {
-        public uint data;
+        public byte data;
         public override string Type() { return "std_msgs/UInt8"; }
         public UInt8()
         {
