+++ Assets/RBSocket/Message/DefaultMsgs/turtlesim/Color.cs
@@ -5,9 +5,9 @@ namespace RBS.Messages.turtlesim
     [System.Serializable]
     public class Color : ExtendMessage
     {
-        public uint r;
-        public uint g;
-        public uint b;
+        public byte r;
+        public byte g;
+        public byte b;
         public override string Type() { return "turtlesim/Color"; }
         public Color()
         {
