+++ Assets/RBSocket/Message/DefaultMsgs/visualization_msgs/InteractiveMarkerFeedback.cs
@@ -5,17 +5,17 @@ namespace RBS.Messages.visualization_msgs
     [System.Serializable]
     public class InteractiveMarkerFeedback : ExtendMessage
     {
-        public uint KEEP_ALIVE;
-        public uint POSE_UPDATE;
-        public uint MENU_SELECT;
-        public uint BUTTON_CLICK;
-        public uint MOUSE_DOWN;
-        public uint MOUSE_UP;
+        public byte KEEP_ALIVE;
+        public byte POSE_UPDATE;
+        public byte MENU_SELECT;
+        public byte BUTTON_CLICK;
+        public byte MOUSE_DOWN;
+        public byte MOUSE_UP;
         public RBS.Messages.std_msgs.Header header;
         public string client_id;
         public string marker_name;
         public string control_name;
-        public uint event_type;
+        public byte event_type;
         public RBS.Messages.geometry_msgs.Pose pose;
         public uint menu_entry_id;
         public RBS.Messages.geometry_msgs.Point mouse_point;
